refactor(application): organize media workflows by domain
