+++ test/unit/application/backgrounds/__init__.py
@@ -1,0 +1,1 @@
+"""Background application tests."""
