+++ test/unit/application/characters/__init__.py
@@ -1,0 +1,1 @@
+"""Character application tests."""
